Fixed produceAchTransmissionInternal() method in RefundServiceImpl class
git-svn-id: https://svn.kuali.org/repos/student/contrib/ksa/trunk@59133 0ac02d80-fb3f-11dc-a574-c57244a850eb

+++ ksa-core/src/main/java/com/sigmasys/kuali/ksa/service/impl/RefundServiceImpl.java
@@ -1327,7 +1327,7 @@ public class RefundServiceImpl extends GenericPersistenceService implements Refu
     private Ach produceAchTransmissionInternal(BigDecimal amount, String reference, Ach ach) {
 
         // Validate the Amount:
-        if (amount == null || amount.intValue() <= 0) {
+        if (amount == null || amount.doubleValue() <= 0) {
             String errMsg = "ACH transmission amount must be a positive number";
             logger.error(errMsg);
             throw new InvalidAchInformationException(errMsg);
